packages/llm-workers/pyproject.toml
[project]
name = "llm-workers"
version = "1.0.0-rc6"
description = "Core library for experimenting with LLMs - provides configuration, workers, tools, and LLM integrations"
authors = [
    {name = "Dmitry Mikhaylov", email = "mikhailov.dmitry@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "pydantic~=2.10.5",
    "python-dotenv~=1.0.1",
    "beautifulsoup4~=4.12.3",
    "PyYAML~=6.0.2",
    "html-text (>=0.7.0,<0.8.0)",
    "RestrictedPython (>=6.0)",
    "langchain (>=1.0.0,<2.0.0)",
    "langchain-core (>=1.0.0,<2.0.0)",
    "langchain-community (>=0.4.0,<2.0.0)",
    "langchain-openai (>=1.0.0,<2.0.0)",
    "langchain-anthropic (>=1.0.0,<2.0.0)",
    "langchain-google-genai (>=4.0.0,<5.0.0)",
    "langchain-mcp-adapters (>=0.1.0,<0.2.0)",
]

[project.optional-dependencies]
html2text = [
    "html2text (>=2025.4.15)",
]

[project.urls]
Homepage = "https://mrbagheera.github.io/llm-workers/"
Issues = "https://github.com/MrBagheera/llm-workers/issues"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "llm_workers", from = "src"}]
